This code is incredibly sloppy.  Just because this is a test-tool is no good excuse to write bad code.  Please read the Java coding conventions as well as the Comoyo code standard.  Also, even though this is a test-tool it should still have unit test coverage.
* Split up file into three.  Please do not put multiple classes in the same file. Follow the
  Java coding conventions.
  http://www.oracle.com/technetwork/java/javase/documentation/codeconventions-141855.html#11684

* Do not use System.out.println for log messages.  Use the logging system that comes
  with the java platform java.util.logging.  Please familiarize yourself with its
  canonical use within Comoyo.

* When writing log messages, please be nice to your fellow human beings and make
  sure that they make sense to someone who didn't write the code.  If a log message
  only makes sense to the author or during development:  delete it before checking
  in the code.

* Add javadoc to classes.  This is a requirement in Comoyo.
